requirements.txt
python-rake
spacy
rake-nltk
git+https://github.com/boudinfl/pke.git
git+https://github.com/LIAAD/yake
numpy

